@@ -1,15 +1,15 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import isEqual from 'lodash/isEqual';
-import get from 'lodash/get';
-import set from 'lodash/set';
-import unset from 'lodash/unset';
-import clone from 'clone';
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import isEqual from "lodash/isEqual";
+import get from "lodash/get";
+import set from "lodash/set";
+import unset from "lodash/unset";
+import clone from "clone";
 
-import bracketsToDots from './shared/bracketsToDots';
-import customPropTypes from './shared/propTypes';
-import filterErrorsForNames from './shared/filterErrorsForNames';
-import recursivelyCloneElements from './shared/recursivelyCloneElements';
+import bracketsToDots from "./shared/bracketsToDots";
+import customPropTypes from "./shared/propTypes";
+import filterErrorsForNames from "./shared/filterErrorsForNames";
+import recursivelyCloneElements from "./shared/recursivelyCloneElements";
 
 // To ensure we do not mutate objects passed in, we'll do a deep clone.
 function cloneValue(value) {
@@ -18,68 +18,6 @@ function cloneValue(value) {
 
 class Form extends Component {
   static isForm = true;
-
-  static propTypes = {
-    children: PropTypes.node.isRequired,
-    className: PropTypes.string,
-    errors: customPropTypes.errors,
-    hasBeenValidated: PropTypes.bool,
-    inputOptions: PropTypes.shape({
-      nullValue: PropTypes.any,
-      propNames: PropTypes.shape({
-        errors: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
-        hasBeenValidated: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
-        isReadOnly: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
-        name: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
-        onChange: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
-        onChanging: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
-        onSubmit: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
-        value: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
-      }),
-    }),
-    logErrorsOnSubmit: PropTypes.bool,
-    // Top-level forms and those under FormList do not need a name
-    name: PropTypes.string, // eslint-disable-line react/no-unused-prop-types
-    onChange: PropTypes.func,
-    onChanging: PropTypes.func,
-    onSubmit: PropTypes.func,
-    revalidateOn: PropTypes.oneOf(['changing', 'changed', 'submit']),
-    style: PropTypes.object, // eslint-disable-line react/forbid-prop-types
-    shouldSubmitWhenInvalid: PropTypes.bool,
-    validateOn: PropTypes.oneOf(['changing', 'changed', 'submit']),
-    validator: PropTypes.func,
-    value: PropTypes.object, // eslint-disable-line react/forbid-prop-types
-  };
-
-  static defaultProps = {
-    className: null,
-    errors: undefined,
-    hasBeenValidated: false,
-    inputOptions: {
-      nullValue: undefined,
-      propNames: {
-        errors: 'errors',
-        hasBeenValidated: 'hasBeenValidated',
-        isReadOnly: 'isReadOnly',
-        name: 'name',
-        onChange: 'onChange',
-        onChanging: 'onChanging',
-        onSubmit: 'onSubmit',
-        value: 'value',
-      },
-    },
-    logErrorsOnSubmit: false,
-    name: null,
-    onChange() {},
-    onChanging() {},
-    onSubmit() {},
-    revalidateOn: 'changing',
-    style: {},
-    shouldSubmitWhenInvalid: false,
-    validateOn: 'submit',
-    validator: undefined,
-    value: undefined,
-  };
 
   constructor(props) {
     super(props);
@@ -97,9 +35,13 @@ class Form extends Component {
     this._isMounted = true;
   }
 
-  componentWillReceiveProps(nextProps) {
+  // eslint-disable-next-line camelcase
+  UNSAFE_componentWillReceiveProps(nextProps) {
     const { hasBeenValidated, value } = this.props;
-    const { hasBeenValidated: hasBeenValidatedNext, value: nextValue } = nextProps;
+    const {
+      hasBeenValidated: hasBeenValidatedNext,
+      value: nextValue,
+    } = nextProps;
 
     // Whenever a changed value prop comes in, we reset state to that, thus becoming clean.
     if (!isEqual(value, nextValue)) {
@@ -107,7 +49,10 @@ class Form extends Component {
     }
 
     // Let props override the `hasBeenValidated` state
-    if (typeof hasBeenValidatedNext === 'boolean' && hasBeenValidatedNext !== hasBeenValidated) {
+    if (
+      typeof hasBeenValidatedNext === "boolean" &&
+      hasBeenValidatedNext !== hasBeenValidated
+    ) {
       this.setState({ hasBeenValidated: hasBeenValidatedNext });
     }
   }
@@ -133,12 +78,10 @@ class Form extends Component {
       this.doSet(this.state.value, fieldName, value);
 
       if (
-        validateOn === 'changed'
-        || validateOn === 'changing'
-        || (
-          hasBeenValidated
-          && (revalidateOn === 'changed' || revalidateOn === 'changing')
-        )
+        validateOn === "changed" ||
+        validateOn === "changing" ||
+        (hasBeenValidated &&
+          (revalidateOn === "changed" || revalidateOn === "changing"))
       ) {
         this.validate().then((updatedErrors) => {
           if (!this._isMounted) return null;
@@ -160,8 +103,8 @@ class Form extends Component {
       this.doSet(this.state.value, fieldName, value);
 
       if (
-        validateOn === 'changing'
-        || (hasBeenValidated && revalidateOn === 'changing')
+        validateOn === "changing" ||
+        (hasBeenValidated && revalidateOn === "changing")
       ) {
         this.validate().then((updatedErrors) => {
           if (!this._isMounted) return null;
@@ -181,7 +124,9 @@ class Form extends Component {
     const { errors } = this.props;
 
     if (!Array.isArray(fieldPaths)) {
-      throw new Error('First argument to getErrors must be an array of field paths');
+      throw new Error(
+        "First argument to getErrors must be an array of field paths"
+      );
     }
 
     return filterErrorsForNames(errors, fieldPaths, !includeDescendantErrors);
@@ -199,15 +144,19 @@ class Form extends Component {
   }
 
   resetValue() {
-    this.setState({
-      errors: [],
-      hasBeenValidated: false,
-      value: cloneValue(this.props.value),
-    }, () => {
-      this.elementRefs.forEach((element) => {
-        if (element && typeof element.resetValue === 'function') element.resetValue();
-      });
-    });
+    this.setState(
+      {
+        errors: [],
+        hasBeenValidated: false,
+        value: cloneValue(this.props.value),
+      },
+      () => {
+        this.elementRefs.forEach((element) => {
+          if (element && typeof element.resetValue === "function")
+            element.resetValue();
+        });
+      }
+    );
   }
 
   doSet(obj, path, value, callback) {
@@ -255,7 +204,9 @@ class Form extends Component {
             // Submission result must be an object with `ok` bool prop
             // and optional submission errors
             if (!Array.isArray(submissionErrors)) {
-              throw new Error('onSubmit returned an errors value that is not an array');
+              throw new Error(
+                "onSubmit returned an errors value that is not an array"
+              );
             }
 
             if (this._isMounted) {
@@ -281,11 +232,13 @@ class Form extends Component {
     const { validator } = this.props;
     const { value } = this.state;
 
-    if (typeof validator !== 'function') return Promise.resolve([]);
+    if (typeof validator !== "function") return Promise.resolve([]);
 
     return validator(value).then((errors) => {
       if (!Array.isArray(errors)) {
-        console.error('validator function must return a Promise that resolves with an array');
+        console.error(
+          "validator function must return a Promise that resolves with an array"
+        );
         return [];
       }
 
@@ -303,10 +256,7 @@ class Form extends Component {
 
     const {
       children,
-      inputOptions: {
-        nullValue,
-        propNames,
-      },
+      inputOptions: { nullValue, propNames },
     } = this.props;
 
     let { errors: propErrors } = this.props;
@@ -330,48 +280,75 @@ class Form extends Component {
         const { names } = element.props;
         if (!names) return {};
         if (element.props[propNames.errors] === undefined) {
-          newProps[propNames.errors] = filterErrorsForNames(errors, names, true);
+          newProps[propNames.errors] = filterErrorsForNames(
+            errors,
+            names,
+            true
+          );
         }
-      } else if (element.type.isFormInput || element.type.isForm || element.type.isFormList) {
+      } else if (
+        element.type.isFormInput ||
+        element.type.isForm ||
+        element.type.isFormList
+      ) {
         const name = element.props[propNames.name];
         if (!name) return {};
 
-        newProps[propNames.onChange] = this.getFieldOnChangeHandler(name, element.props[propNames.onChange]);
-        newProps[propNames.onChanging] = this.getFieldOnChangingHandler(name, element.props[propNames.onChanging]);
-        newProps[propNames.onSubmit] = this.getFieldOnSubmitHandler(element.props[propNames.onSubmit]);
+        newProps[propNames.onChange] = this.getFieldOnChangeHandler(
+          name,
+          element.props[propNames.onChange]
+        );
+        newProps[propNames.onChanging] = this.getFieldOnChangingHandler(
+          name,
+          element.props[propNames.onChanging]
+        );
+        newProps[propNames.onSubmit] = this.getFieldOnSubmitHandler(
+          element.props[propNames.onSubmit]
+        );
 
         if (element.props[propNames.value] === undefined) {
           // Some input components (MUI) do not accept a `null` value.
           // For these, passing `{ nullValue: "" }` options does the trick.
           let inputValue = get(value, name);
-          if (inputValue === null && nullValue !== undefined) inputValue = nullValue;
+          if (inputValue === null && nullValue !== undefined)
+            inputValue = nullValue;
           newProps[propNames.value] = inputValue;
         }
 
         if (element.props[propNames.errors] === undefined) {
-          newProps[propNames.errors] = filterErrorsForNames(errors, [name], false);
+          newProps[propNames.errors] = filterErrorsForNames(
+            errors,
+            [name],
+            false
+          );
 
           // Adjust the error names to correct scope
           if (element.type.isForm) {
             const canonicalName = bracketsToDots(name);
-            newProps[propNames.errors] = newProps[propNames.errors].map((err) => {
-              return {
-                ...err,
-                name: bracketsToDots(err.name).slice(canonicalName.length + 1),
-              };
-            });
+            newProps[propNames.errors] = newProps[propNames.errors].map(
+              (err) => {
+                return {
+                  ...err,
+                  name: bracketsToDots(err.name).slice(
+                    canonicalName.length + 1
+                  ),
+                };
+              }
+            );
           }
         }
 
         newProps[propNames.hasBeenValidated] = hasBeenValidated;
 
         if (element.type.isFormInput) {
-          if (typeof element.props[propNames.isReadOnly] === 'function') {
+          if (typeof element.props[propNames.isReadOnly] === "function") {
             newProps[propNames.isReadOnly] = element.props.isReadOnly(value);
           }
         }
 
-        newProps.ref = (el) => { this.elementRefs.push(el); };
+        newProps.ref = (el) => {
+          this.elementRefs.push(el);
+        };
       }
 
       return newProps;
@@ -393,5 +370,68 @@ class Form extends Component {
     );
   }
 }
+
+Form.propTypes = {
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string,
+  errors: customPropTypes.errors,
+  hasBeenValidated: PropTypes.bool,
+  inputOptions: PropTypes.shape({
+    // eslint-disable-next-line react/forbid-prop-types
+    nullValue: PropTypes.any,
+    propNames: PropTypes.shape({
+      errors: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+      hasBeenValidated: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+      isReadOnly: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+      name: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+      onChange: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+      onChanging: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+      onSubmit: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+      value: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+    }),
+  }),
+  logErrorsOnSubmit: PropTypes.bool,
+  // Top-level forms and those under FormList do not need a name
+  name: PropTypes.string, // eslint-disable-line react/no-unused-prop-types
+  onChange: PropTypes.func,
+  onChanging: PropTypes.func,
+  onSubmit: PropTypes.func,
+  revalidateOn: PropTypes.oneOf(["changing", "changed", "submit"]),
+  style: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  shouldSubmitWhenInvalid: PropTypes.bool,
+  validateOn: PropTypes.oneOf(["changing", "changed", "submit"]),
+  validator: PropTypes.func,
+  value: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+};
+
+Form.defaultProps = {
+  className: null,
+  errors: undefined,
+  hasBeenValidated: false,
+  inputOptions: {
+    nullValue: undefined,
+    propNames: {
+      errors: "errors",
+      hasBeenValidated: "hasBeenValidated",
+      isReadOnly: "isReadOnly",
+      name: "name",
+      onChange: "onChange",
+      onChanging: "onChanging",
+      onSubmit: "onSubmit",
+      value: "value",
+    },
+  },
+  logErrorsOnSubmit: false,
+  name: null,
+  onChange() {},
+  onChanging() {},
+  onSubmit() {},
+  revalidateOn: "changing",
+  style: {},
+  shouldSubmitWhenInvalid: false,
+  validateOn: "submit",
+  validator: undefined,
+  value: undefined,
+};
 
 export default Form;
