@@ -1,12 +1,12 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import isEqual from 'lodash/isEqual';
-import clone from 'clone';
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import isEqual from "lodash/isEqual";
+import clone from "clone";
 
-import bracketsToDots from './shared/bracketsToDots';
-import customPropTypes from './shared/propTypes';
-import filterErrorsForNames from './shared/filterErrorsForNames';
-import recursivelyCloneElements from './shared/recursivelyCloneElements';
+import bracketsToDots from "./shared/bracketsToDots";
+import customPropTypes from "./shared/propTypes";
+import filterErrorsForNames from "./shared/filterErrorsForNames";
+import recursivelyCloneElements from "./shared/recursivelyCloneElements";
 
 const styles = {
   button: {
@@ -14,101 +14,53 @@ const styles = {
     paddingRight: 0,
     paddingLeft: 0,
     paddingBottom: 0,
-    width: '3.5rem',
-    height: '3.5rem',
-    verticalAlign: '-webkit-baseline-middle',
+    width: "3.5rem",
+    height: "3.5rem",
+    verticalAlign: "-webkit-baseline-middle",
   },
   item: {
-    display: 'flex',
-    flexDirection: 'row',
+    display: "flex",
+    flexDirection: "row",
     borderWidth: 1,
-    borderColor: '#cccccc',
-    borderBottomStyle: 'solid',
-    borderTopStyle: 'none',
-    borderLeftStyle: 'none',
-    borderRightStyle: 'none',
-    paddingBottom: '1rem',
-    marginBottom: '1rem',
+    borderColor: "#cccccc",
+    borderBottomStyle: "solid",
+    borderTopStyle: "none",
+    borderLeftStyle: "none",
+    borderRightStyle: "none",
+    paddingBottom: "1rem",
+    marginBottom: "1rem",
   },
   itemLeft: {
-    paddingRight: '1.5rem',
+    paddingRight: "1.5rem",
   },
   itemRight: {
     flexGrow: 1,
   },
   lastItem: {
-    display: 'flex',
-    flexDirection: 'row',
+    display: "flex",
+    flexDirection: "row",
   },
   list: {
     borderWidth: 1,
-    borderColor: '#cccccc',
-    borderStyle: 'solid',
+    borderColor: "#cccccc",
+    borderStyle: "solid",
     borderRadius: 3,
-    padding: '1rem',
+    padding: "1rem",
   },
   addItemRow: {
     borderWidth: 1,
-    borderColor: '#cccccc',
-    borderBottomStyle: 'none',
-    borderTopStyle: 'solid',
-    borderLeftStyle: 'none',
-    borderRightStyle: 'none',
-    paddingTop: '1rem',
-    marginTop: '1rem',
+    borderColor: "#cccccc",
+    borderBottomStyle: "none",
+    borderTopStyle: "solid",
+    borderLeftStyle: "none",
+    borderRightStyle: "none",
+    paddingTop: "1rem",
+    marginTop: "1rem",
   },
 };
 
 class FormList extends Component {
   static isFormList = true;
-
-  static propTypes = {
-    addButtonText: PropTypes.string,
-    addItemRowStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
-    buttonClassName: PropTypes.string,
-    buttonStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
-    children: PropTypes.node.isRequired,
-    className: PropTypes.string,
-    errors: customPropTypes.errors,
-    itemAreaClassName: PropTypes.string,
-    itemAreaStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
-    itemClassName: PropTypes.string,
-    itemStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
-    itemRemoveAreaClassName: PropTypes.string,
-    itemRemoveAreaStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
-    maxCount: PropTypes.number,
-    minCount: PropTypes.number,
-    name: PropTypes.string.isRequired,
-    onChanging: PropTypes.func,
-    onChange: PropTypes.func,
-    onSubmit: PropTypes.func,
-    removeButtonText: PropTypes.string,
-    style: PropTypes.object, // eslint-disable-line react/forbid-prop-types
-    value: PropTypes.arrayOf(PropTypes.any),
-  };
-
-  static defaultProps = {
-    addButtonText: '+',
-    addItemRowStyle: {},
-    buttonClassName: null,
-    buttonStyle: {},
-    className: null,
-    errors: undefined,
-    itemAreaClassName: null,
-    itemAreaStyle: {},
-    itemClassName: null,
-    itemStyle: {},
-    itemRemoveAreaClassName: null,
-    itemRemoveAreaStyle: {},
-    minCount: 0,
-    maxCount: undefined,
-    onChanging() {},
-    onChange() {},
-    onSubmit() {},
-    removeButtonText: '–',
-    style: {},
-    value: undefined,
-  };
 
   constructor(props) {
     super(props);
@@ -125,13 +77,15 @@ class FormList extends Component {
     this.elementRefs = [];
   }
 
-  componentWillMount() {
+  // eslint-disable-next-line camelcase
+  UNSAFE_componentWillMount() {
     const { value } = this.state;
     this.handleChanging(value);
     this.handleChanged(value);
   }
 
-  componentWillReceiveProps(nextProps) {
+  // eslint-disable-next-line camelcase
+  UNSAFE_componentWillReceiveProps(nextProps) {
     const { value } = this.props;
     let { minCount, value: nextValue } = nextProps;
     const { value: stateValue } = this.state;
@@ -154,6 +108,36 @@ class FormList extends Component {
       this.handleChanging(nextValue);
       this.handleChanged(nextValue);
     }
+  }
+
+  handleChanged(value) {
+    const { onChange } = this.props;
+    if (!isEqual(value, this.lastChangedValue)) {
+      this.lastChangedValue = value;
+      onChange(value);
+    }
+  }
+
+  handleChanging(value) {
+    const { onChanging } = this.props;
+    if (!isEqual(value, this.lastChangingValue)) {
+      this.lastChangingValue = value;
+      onChanging(value);
+    }
+  }
+
+  handleClickRemoveItem(index) {
+    return () => {
+      const { minCount } = this.props;
+      const { value } = this.state;
+
+      if (value.length === Math.max(minCount || 0, 0)) return;
+
+      value.splice(index, 1); // mutation is ok because we cloned, and likely faster
+      this.setState({ value });
+      this.handleChanging(value);
+      this.handleChanged(value);
+    };
   }
 
   getFieldValueHandler(index, isChanged) {
@@ -179,22 +163,6 @@ class FormList extends Component {
     this.setState({ value });
   };
 
-  handleChanging(value) {
-    const { onChanging } = this.props;
-    if (!isEqual(value, this.lastChangingValue)) {
-      this.lastChangingValue = value;
-      onChanging(value);
-    }
-  }
-
-  handleChanged(value) {
-    const { onChange } = this.props;
-    if (!isEqual(value, this.lastChangedValue)) {
-      this.lastChangedValue = value;
-      onChange(value);
-    }
-  }
-
   resetValue() {
     let { minCount, value } = this.props;
     value = clone(value || []);
@@ -205,22 +173,9 @@ class FormList extends Component {
     this.setState({ value });
 
     this.elementRefs.forEach((element) => {
-      if (element && typeof element.resetValue === 'function') element.resetValue();
+      if (element && typeof element.resetValue === "function")
+        element.resetValue();
     });
-  }
-
-  handleClickRemoveItem(index) {
-    return () => {
-      const { minCount } = this.props;
-      const { value } = this.state;
-
-      if (value.length === Math.max(minCount || 0, 0)) return;
-
-      value.splice(index, 1); // mutation is ok because we cloned, and likely faster
-      this.setState({ value });
-      this.handleChanging(value);
-      this.handleChanged(value);
-    };
   }
 
   renderArrayItems() {
@@ -249,14 +204,22 @@ class FormList extends Component {
     let errorsChild;
     React.Children.forEach(children, (child) => {
       if (child.type.isFormList) {
-        throw new Error('reacto-form FormList: FormList may not be a child of FormList');
+        throw new Error(
+          "reacto-form FormList: FormList may not be a child of FormList"
+        );
       }
       if (child.type.isForm || child.type.isFormInput) {
-        if (itemChild) throw new Error('reacto-form FormList: FormList must have exactly one Input or Form child');
+        if (itemChild)
+          throw new Error(
+            "reacto-form FormList: FormList must have exactly one Input or Form child"
+          );
         itemChild = child;
       }
       if (child.type.isFormErrors) {
-        if (errorsChild) throw new Error('reacto-form FormList: FormList may have no more than one ErrorsBlock child');
+        if (errorsChild)
+          throw new Error(
+            "reacto-form FormList: FormList may have no more than one ErrorsBlock child"
+          );
         errorsChild = child;
       }
     });
@@ -275,43 +238,69 @@ class FormList extends Component {
             filteredErrors = filteredErrors.map((err) => {
               return {
                 ...err,
-                name: bracketsToDots(err.name).slice(`${name}.${index}`.length + 1),
+                name: bracketsToDots(err.name).slice(
+                  `${name}.${index}`.length + 1
+                ),
               };
             });
           }
 
-          return React.cloneElement(child, {
-            errors: filteredErrors,
-            name: itemName,
-            onChange: this.getFieldValueHandler(index, true),
-            onChanging: this.getFieldValueHandler(index, false),
-            onSubmit,
-            ref: (el) => { this.elementRefs.push(el); },
-            value: itemValue,
-          }, recursivelyCloneElements(child.props.children));
+          return React.cloneElement(
+            child,
+            {
+              errors: filteredErrors,
+              name: itemName,
+              onChange: this.getFieldValueHandler(index, true),
+              onChanging: this.getFieldValueHandler(index, false),
+              onSubmit,
+              ref: (el) => {
+                this.elementRefs.push(el);
+              },
+              value: itemValue,
+            },
+            recursivelyCloneElements(child.props.children)
+          );
         }
 
         if (child.type.isFormErrors) {
-          return React.cloneElement(child, {
-            errors: filterErrorsForNames(errors, [itemName], true),
-            names: [itemName],
-          }, recursivelyCloneElements(child.props.children));
+          return React.cloneElement(
+            child,
+            {
+              errors: filterErrorsForNames(errors, [itemName], true),
+              names: [itemName],
+            },
+            recursivelyCloneElements(child.props.children)
+          );
         }
 
         return recursivelyCloneElements(child);
       });
 
-      let resolvedItemStyle = index + 1 === value.length ? styles.lastItem : styles.item;
+      let resolvedItemStyle =
+        index + 1 === value.length ? styles.lastItem : styles.item;
       resolvedItemStyle = { ...resolvedItemStyle, ...itemStyle };
 
       return (
         <div className={itemClassName} key={itemName} style={resolvedItemStyle}>
           {hasMoreThanMinCount && (
-            <div className={itemRemoveAreaClassName} style={{ ...styles.itemLeft, ...itemRemoveAreaStyle }}>
-              <button type="button" className={buttonClassName} onClick={this.handleClickRemoveItem(index)} style={{ ...styles.button, ...buttonStyle }}>{removeButtonText}</button>
+            <div
+              className={itemRemoveAreaClassName}
+              style={{ ...styles.itemLeft, ...itemRemoveAreaStyle }}
+            >
+              <button
+                type="button"
+                className={buttonClassName}
+                onClick={this.handleClickRemoveItem(index)}
+                style={{ ...styles.button, ...buttonStyle }}
+              >
+                {removeButtonText}
+              </button>
             </div>
           )}
-          <div className={itemAreaClassName} style={{ ...styles.itemRight, ...itemAreaStyle }}>
+          <div
+            className={itemAreaClassName}
+            style={{ ...styles.itemRight, ...itemAreaStyle }}
+          >
             {kids}
           </div>
         </div>
@@ -334,7 +323,14 @@ class FormList extends Component {
 
     return (
       <div className={itemClassName} style={resolvedStyle}>
-        <button type="button" className={buttonClassName} onClick={this.handleClickAddItem} style={styles.button}>{addButtonText}</button>
+        <button
+          type="button"
+          className={buttonClassName}
+          onClick={this.handleClickAddItem}
+          style={styles.button}
+        >
+          {addButtonText}
+        </button>
       </div>
     );
   }
@@ -343,7 +339,8 @@ class FormList extends Component {
     const { className, maxCount, style } = this.props;
     let { value } = this.state;
     if (!value) value = [];
-    const hasFewerThanMaxCount = value.length < maxCount || maxCount === undefined || maxCount === null;
+    const hasFewerThanMaxCount =
+      value.length < maxCount || maxCount === undefined || maxCount === null;
 
     return (
       <div className={className} style={{ ...styles.list, ...style }}>
@@ -353,5 +350,53 @@ class FormList extends Component {
     );
   }
 }
+
+FormList.propTypes = {
+  addButtonText: PropTypes.string,
+  addItemRowStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  buttonClassName: PropTypes.string,
+  buttonStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string,
+  errors: customPropTypes.errors,
+  itemAreaClassName: PropTypes.string,
+  itemAreaStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  itemClassName: PropTypes.string,
+  itemStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  itemRemoveAreaClassName: PropTypes.string,
+  itemRemoveAreaStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  maxCount: PropTypes.number,
+  minCount: PropTypes.number,
+  name: PropTypes.string.isRequired,
+  onChanging: PropTypes.func,
+  onChange: PropTypes.func,
+  onSubmit: PropTypes.func,
+  removeButtonText: PropTypes.string,
+  style: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  value: PropTypes.arrayOf(PropTypes.any),
+};
+
+FormList.defaultProps = {
+  addButtonText: "+",
+  addItemRowStyle: {},
+  buttonClassName: null,
+  buttonStyle: {},
+  className: null,
+  errors: undefined,
+  itemAreaClassName: null,
+  itemAreaStyle: {},
+  itemClassName: null,
+  itemStyle: {},
+  itemRemoveAreaClassName: null,
+  itemRemoveAreaStyle: {},
+  minCount: 0,
+  maxCount: undefined,
+  onChanging() {},
+  onChange() {},
+  onSubmit() {},
+  removeButtonText: "–",
+  style: {},
+  value: undefined,
+};
 
 export default FormList;
