@@ -1,31 +1,31 @@
-import React from 'react';
-import { mount } from 'enzyme';
-import renderer from 'react-test-renderer';
-import { Field, Input } from 'reacto-form-inputs';
-import Form from './Form';
+import React from "react";
+import { mount } from "enzyme";
+import renderer from "react-test-renderer";
+import { Field, Input } from "reacto-form-inputs";
+import Form from "./Form";
 
-test('has isForm property set to true', () => {
+test("has isForm property set to true", () => {
   expect(Form.isForm).toBe(true);
 });
 
-test('form snapshot 1', () => {
+test("form snapshot 1", () => {
   const component = renderer.create(
-    <Form value={{ foo: { bar: 'TEST VALUE' } }}>
+    <Form value={{ foo: { bar: "TEST VALUE" } }}>
       <Field label="LABEL">
         <p>Text above</p>
         <Input name="foo.bar" />
         <p>Text below</p>
       </Field>
-    </Form>,
+    </Form>
   );
 
   const tree = component.toJSON();
   expect(tree).toMatchSnapshot();
 });
 
-test('form snapshot 2 - complex nesting', () => {
+test("form snapshot 2 - complex nesting", () => {
   const component = renderer.create(
-    <Form value={{ foo: { bar: 'TEST VALUE' }, type: 'monkey' }}>
+    <Form value={{ foo: { bar: "TEST VALUE" }, type: "monkey" }}>
       <Field label="TYPE LABEL">
         <p>Text above</p>
         <Input name="type" />
@@ -46,46 +46,46 @@ test('form snapshot 2 - complex nesting', () => {
           </Field>
         </Form>
       </div>
-    </Form>,
+    </Form>
   );
 
   const tree = component.toJSON();
   expect(tree).toMatchSnapshot();
 });
 
-test('sets value prop on child input for simple name', () => {
+test("sets value prop on child input for simple name", () => {
   const wrapper = mount(
-    <Form value={{ foo: 'BAR' }}>
+    <Form value={{ foo: "BAR" }}>
       <Input name="foo" />
-    </Form>,
+    </Form>
   );
 
-  expect(wrapper.find('input').prop('value')).toBe('BAR');
+  expect(wrapper.find("input").prop("value")).toBe("BAR");
 });
 
-test('sets value prop on child input for path name', () => {
+test("sets value prop on child input for path name", () => {
   const wrapper = mount(
-    <Form value={{ foo: [{ a: 'VAL' }] }}>
+    <Form value={{ foo: [{ a: "VAL" }] }}>
       <Input name="foo[0].a" />
-    </Form>,
+    </Form>
   );
 
-  expect(wrapper.find('input').prop('value')).toBe('VAL');
+  expect(wrapper.find("input").prop("value")).toBe("VAL");
 });
 
-test('keeps child input value prop if present', () => {
+test("keeps child input value prop if present", () => {
   const wrapper = mount(
-    <Form value={{ foo: 'BAR' }}>
+    <Form value={{ foo: "BAR" }}>
       <Input name="foo" value="DEFAULT" />
-    </Form>,
+    </Form>
   );
 
-  expect(wrapper.find('input').prop('value')).toBe('DEFAULT');
+  expect(wrapper.find("input").prop("value")).toBe("DEFAULT");
 });
 
-test('sets value prop on nested descendant input', () => {
+test("sets value prop on nested descendant input", () => {
   const wrapper = mount(
-    <Form value={{ foo: 'BAR' }}>
+    <Form value={{ foo: "BAR" }}>
       <div>
         <Field label="LABEL">
           <div>
@@ -93,48 +93,48 @@ test('sets value prop on nested descendant input', () => {
           </div>
         </Field>
       </div>
-    </Form>,
+    </Form>
   );
 
-  expect(wrapper.find('input').prop('value')).toBe('BAR');
+  expect(wrapper.find("input").prop("value")).toBe("BAR");
 });
 
-test('simple form value is updated after user enters input', () => {
+test("simple form value is updated after user enters input", () => {
   const wrapper = mount(
-    <Form value={{ foo: 'BAR' }}>
+    <Form value={{ foo: "BAR" }}>
       <Input name="foo" />
-    </Form>,
+    </Form>
   );
 
-  expect(wrapper.instance().getValue()).toEqual({ foo: 'BAR' });
+  expect(wrapper.instance().getValue()).toEqual({ foo: "BAR" });
 
-  wrapper.find('input').simulate('change', { target: { value: 'NEW' } });
+  wrapper.find("input").simulate("change", { target: { value: "NEW" } });
 
-  expect(wrapper.instance().getValue()).toEqual({ foo: 'NEW' });
+  expect(wrapper.instance().getValue()).toEqual({ foo: "NEW" });
 });
 
-test('path form value is updated after user enters input', () => {
+test("path form value is updated after user enters input", () => {
   const wrapper = mount(
-    <Form value={{ foo: [{ a: 'VAL' }] }}>
+    <Form value={{ foo: [{ a: "VAL" }] }}>
       <Input name="foo[0].a" />
-    </Form>,
+    </Form>
   );
 
-  expect(wrapper.instance().getValue()).toEqual({ foo: [{ a: 'VAL' }] });
+  expect(wrapper.instance().getValue()).toEqual({ foo: [{ a: "VAL" }] });
 
-  wrapper.find('input').simulate('change', { target: { value: 'NEW' } });
+  wrapper.find("input").simulate("change", { target: { value: "NEW" } });
 
-  expect(wrapper.instance().getValue()).toEqual({ foo: [{ a: 'NEW' }] });
+  expect(wrapper.instance().getValue()).toEqual({ foo: [{ a: "NEW" }] });
 });
 
-test('blurring input triggers form onChanging and onChange', () => {
-  const onChange = jest.fn().mockName('onChange');
-  const onChanging = jest.fn().mockName('onChanging');
+test("blurring input triggers form onChanging and onChange", () => {
+  const onChange = jest.fn().mockName("onChange");
+  const onChanging = jest.fn().mockName("onChanging");
 
   const wrapper = mount(
     <Form onChange={onChange} onChanging={onChanging}>
       <Input name="foo" />
-    </Form>,
+    </Form>
   );
 
   expect(onChange).toHaveBeenCalledTimes(1);
@@ -146,11 +146,11 @@ test('blurring input triggers form onChanging and onChange', () => {
   onChange.mockClear();
   onChanging.mockClear();
 
-  wrapper.find('input').simulate('blur', { target: { value: 'NEW' } });
+  wrapper.find("input").simulate("blur", { target: { value: "NEW" } });
 
   expect(onChange).toHaveBeenCalledTimes(1);
   expect(onChanging).toHaveBeenCalledTimes(1);
 
-  expect(onChange.mock.calls[0][0]).toEqual({ foo: 'NEW' });
-  expect(onChanging.mock.calls[0][0]).toEqual({ foo: 'NEW' });
+  expect(onChange.mock.calls[0][0]).toEqual({ foo: "NEW" });
+  expect(onChanging.mock.calls[0][0]).toEqual({ foo: "NEW" });
 });
