@@ -1,12 +1,7 @@
 /* eslint-disable import/no-named-as-default,import/no-named-as-default-member */
-import Form from './Form';
-import FormList from './FormList';
-import muiOptions from './muiOptions';
-import useReactoForm from './useReactoForm';
+import Form from "./Form";
+import FormList from "./FormList";
+import muiOptions from "./muiOptions";
+import useReactoForm from "./useReactoForm";
 
-export {
-  Form,
-  FormList,
-  muiOptions,
-  useReactoForm,
-};
+export { Form, FormList, muiOptions, useReactoForm };
