@@ -1,12 +1,12 @@
 export default {
   nullValue: false,
-  onChangeGetValue: event => (event.target.checked || false),
+  onChangeGetValue: (event) => event.target.checked || false,
   propNames: {
     errors: false,
     hasBeenValidated: false,
-    isReadOnly: 'disabled',
+    isReadOnly: "disabled",
     onChanging: false,
     onSubmit: false,
-    value: 'checked',
+    value: "checked",
   },
 };
