@@ -1,13 +1,13 @@
 export default {
-  nullValue: '',
-  onChangeGetValue: event => event.target.value,
-  onChangingGetValue: event => event.target.value,
+  nullValue: "",
+  onChangeGetValue: (event) => event.target.value,
+  onChangingGetValue: (event) => event.target.value,
   propNames: {
     errors: false,
     hasBeenValidated: false,
-    isReadOnly: 'disabled',
-    onChange: 'onBlur',
-    onChanging: 'onChange',
+    isReadOnly: "disabled",
+    onChange: "onBlur",
+    onChanging: "onChange",
     onSubmit: false,
   },
 };
