@@ -1,5 +1,5 @@
-import toPath from 'lodash/toPath';
+import toPath from "lodash/toPath";
 
 export default function bracketsToDots(pathString) {
-  return toPath(pathString).join('.');
+  return toPath(pathString).join(".");
 }
