@@ -1,16 +1,16 @@
-import bracketsToDots from './bracketsToDots';
+import bracketsToDots from "./bracketsToDots";
 
 export default function filterErrorsForNames(errors, names, exact) {
   if (!Array.isArray(errors) || !Array.isArray(names)) return [];
 
   // Accept paths that may contain brackets or dots, making them all dots
-  names = names.map(name => bracketsToDots(name));
+  names = names.map((name) => bracketsToDots(name));
 
   return errors.filter((error) => {
     const errorName = bracketsToDots(error.name);
     return names.some((name) => {
       if (name === errorName) return true;
-      return (!exact && errorName.startsWith(`${name}.`));
+      return !exact && errorName.startsWith(`${name}.`);
     });
   });
 }
