@@ -6,20 +6,25 @@
  * @param {String} timezone - A timezone identifier
  */
 export default function getDateTimeValuesFromDate(date, moment, timezone) {
-  if (!date || !(date instanceof Date) || !moment || typeof timezone !== 'string') {
+  if (
+    !date ||
+    !(date instanceof Date) ||
+    !moment ||
+    typeof timezone !== "string"
+  ) {
     // The defaults have to be "" rather than undefined or null so that React knows they are "controlled" inputs
     return {
-      dayValue: '',
-      monthValue: '',
-      timeValue: '',
-      yearValue: '',
+      dayValue: "",
+      monthValue: "",
+      timeValue: "",
+      yearValue: "",
     };
   }
   const m = moment(date).tz(timezone);
   return {
-    dayValue: m.format('DD'),
-    monthValue: m.format('M'),
-    timeValue: m.format('HH:mm'),
-    yearValue: m.format('YYYY'),
+    dayValue: m.format("DD"),
+    monthValue: m.format("M"),
+    timeValue: m.format("HH:mm"),
+    yearValue: m.format("YYYY"),
   };
 }
