@@ -1,10 +1,12 @@
-import PropTypes from 'prop-types';
+import PropTypes from "prop-types";
 
 const customPropTypes = {
-  errors: PropTypes.arrayOf(PropTypes.shape({
-    message: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-  })),
+  errors: PropTypes.arrayOf(
+    PropTypes.shape({
+      message: PropTypes.string.isRequired,
+      name: PropTypes.string.isRequired,
+    })
+  ),
 };
 
 export default customPropTypes;
