@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 
 /**
  * Recursively clone elements
@@ -9,17 +9,27 @@ import React from 'react';
  * @param {Function} [shouldStopRecursing] - A function that takes an element and returns a truthy value if `recursivelyCloneElements`
  *                                         should not be called for that element's children
  */
-export default function recursivelyCloneElements(elements, getProps, shouldStopRecursing) {
+export default function recursivelyCloneElements(
+  elements,
+  getProps,
+  shouldStopRecursing
+) {
   const newElements = React.Children.map(elements, (element) => {
-    if (!element || typeof element === 'string' || !element.props) return element;
+    if (!element || typeof element === "string" || !element.props)
+      return element;
 
-    if (typeof getProps !== 'function') getProps = () => ({});
-    if (typeof shouldStopRecursing !== 'function') shouldStopRecursing = () => false;
+    if (typeof getProps !== "function") getProps = () => ({});
+    if (typeof shouldStopRecursing !== "function")
+      shouldStopRecursing = () => false;
 
-    const children = shouldStopRecursing(element) ? element.props.children : recursivelyCloneElements(element.props.children, getProps);
+    const children = shouldStopRecursing(element)
+      ? element.props.children
+      : recursivelyCloneElements(element.props.children, getProps);
 
     return React.cloneElement(element, getProps(element), children);
   });
 
-  return Array.isArray(newElements) && newElements.length === 1 ? newElements[0] : newElements;
+  return Array.isArray(newElements) && newElements.length === 1
+    ? newElements[0]
+    : newElements;
 }
