@@ -1,10 +1,10 @@
-import { useEffect, useState } from 'react';
-import clone from 'clone';
-import get from 'lodash/get';
-import isEqual from 'lodash/isEqual';
-import set from 'lodash/set';
-import bracketsToDots from './shared/bracketsToDots';
-import filterErrorsForNames from './shared/filterErrorsForNames';
+import { useEffect, useState } from "react";
+import clone from "clone";
+import get from "lodash/get";
+import isEqual from "lodash/isEqual";
+import set from "lodash/set";
+import bracketsToDots from "./shared/bracketsToDots";
+import filterErrorsForNames from "./shared/filterErrorsForNames";
 
 /**
  * @summary To ensure we do not mutate objects passed in, we'll do a deep clone.
@@ -25,22 +25,24 @@ export default function useReactoForm(props) {
     onChange = () => {},
     onChanging = () => {},
     onSubmit = () => {},
-    revalidateOn = 'changing',
+    revalidateOn = "changing",
     shouldSubmitWhenInvalid = false,
-    validateOn = 'submit',
+    validateOn = "submit",
     validator,
     value: valueProp,
   } = props;
 
   const [errors, setErrors] = useState([]);
-  const [hasBeenValidated, setHasBeenValidated] = useState(hasBeenValidatedProp || false);
+  const [hasBeenValidated, setHasBeenValidated] = useState(
+    hasBeenValidatedProp || false
+  );
   const [forceReset, setForceReset] = useState(0);
   const [formData, setFormData] = useState({});
 
   // isReadOnly can be passed as a function, which is then called with
   // the current form data to determine whether it should be read only.
   let { isReadOnly } = props;
-  if (typeof isReadOnly === 'function') {
+  if (typeof isReadOnly === "function") {
     isReadOnly = !!isReadOnly(formData);
   }
 
@@ -50,7 +52,7 @@ export default function useReactoForm(props) {
    */
   function setFieldValueInFormData(fieldPath, fieldValue) {
     const formDataCopy = clone(formData);
-    set(formDataCopy, fieldPath, (fieldValue === undefined ? null : fieldValue));
+    set(formDataCopy, fieldPath, fieldValue === undefined ? null : fieldValue);
     setFormData(formDataCopy);
     return formDataCopy;
   }
@@ -64,7 +66,10 @@ export default function useReactoForm(props) {
 
   // Let props override the `hasBeenValidated` state
   useEffect(() => {
-    if (typeof hasBeenValidatedProp === 'boolean' && hasBeenValidatedProp !== hasBeenValidated) {
+    if (
+      typeof hasBeenValidatedProp === "boolean" &&
+      hasBeenValidatedProp !== hasBeenValidated
+    ) {
       setHasBeenValidated(hasBeenValidatedProp);
     }
   }, [hasBeenValidatedProp]);
@@ -74,7 +79,7 @@ export default function useReactoForm(props) {
    * @return {Promise<Object[]>} Promised array of error objects
    */
   async function validateForm() {
-    if (typeof validator !== 'function') return [];
+    if (typeof validator !== "function") return [];
 
     let customValidatorErrors;
     try {
@@ -85,7 +90,7 @@ export default function useReactoForm(props) {
     }
 
     if (!Array.isArray(customValidatorErrors)) {
-      console.error('validator function must return or promise an array');
+      console.error("validator function must return or promise an array");
       return [];
     }
 
@@ -101,7 +106,8 @@ export default function useReactoForm(props) {
   async function submitForm() {
     const validationErrors = await validateForm();
 
-    if (logErrorsOnSubmit && validationErrors.length > 0) console.error(validationErrors);
+    if (logErrorsOnSubmit && validationErrors.length > 0)
+      console.error(validationErrors);
 
     if (validationErrors.length && !shouldSubmitWhenInvalid) return null;
 
@@ -109,7 +115,10 @@ export default function useReactoForm(props) {
     // for submission to complete, but we won't worry about it if it doesn't
     let submissionResult;
     try {
-      submissionResult = await onSubmit(formData, validationErrors.length === 0);
+      submissionResult = await onSubmit(
+        formData,
+        validationErrors.length === 0
+      );
     } catch (error) {
       console.error('Form "onSubmit" function error:', error);
       return null;
@@ -130,7 +139,9 @@ export default function useReactoForm(props) {
       if (Array.isArray(submissionErrors)) {
         setErrors(submissionErrors);
       } else {
-        console.error('onSubmit returned an object with "errors" property that is not a valid errors array');
+        console.error(
+          'onSubmit returned an object with "errors" property that is not a valid errors array'
+        );
       }
     }
 
@@ -139,7 +150,9 @@ export default function useReactoForm(props) {
 
   function getErrors(fieldPaths, { includeDescendantErrors = false } = {}) {
     if (!Array.isArray(fieldPaths)) {
-      throw new Error('First argument to getErrors must be an array of field paths');
+      throw new Error(
+        "First argument to getErrors must be an array of field paths"
+      );
     }
     return filterErrorsForNames(errors, fieldPaths, !includeDescendantErrors);
   }
@@ -166,7 +179,7 @@ export default function useReactoForm(props) {
       // Adjust the error names to correct scope for forms
       if (isForm) {
         const canonicalName = bracketsToDots(fieldPath);
-        fieldErrors = fieldErrors.map(err => ({
+        fieldErrors = fieldErrors.map((err) => ({
           ...err,
           name: bracketsToDots(err.name).slice(canonicalName.length + 1),
         }));
@@ -177,18 +190,18 @@ export default function useReactoForm(props) {
         // directly as the first arg. Many popular libraries pass
         // an Event as the first arg, and `onChangeGetValue` can be
         // used to determine and return the new value.
-        const inputValue = onChangeGetValue ? onChangeGetValue(...onChangeArgs) : onChangeArgs[0];
+        const inputValue = onChangeGetValue
+          ? onChangeGetValue(...onChangeArgs)
+          : onChangeArgs[0];
 
         const updatedFormData = setFieldValueInFormData(fieldPath, inputValue);
 
         // Now bubble up the `onChange`, possibly validating first
         if (
-          validateOn === 'changed'
-          || validateOn === 'changing'
-          || (
-            hasBeenValidated
-            && (revalidateOn === 'changed' || revalidateOn === 'changing')
-          )
+          validateOn === "changed" ||
+          validateOn === "changing" ||
+          (hasBeenValidated &&
+            (revalidateOn === "changed" || revalidateOn === "changing"))
         ) {
           validateForm()
             .then((updatedErrors) => {
@@ -208,13 +221,15 @@ export default function useReactoForm(props) {
         // directly as the first arg. Many popular libraries pass
         // an Event as the first arg, and `onChangeGetValue` can be
         // used to determine and return the new value.
-        const inputValue = onChangingGetValue ? onChangingGetValue(...onChangingArgs) : onChangingArgs[0];
+        const inputValue = onChangingGetValue
+          ? onChangingGetValue(...onChangingArgs)
+          : onChangingArgs[0];
 
         const updatedFormData = setFieldValueInFormData(fieldPath, inputValue);
 
         if (
-          validateOn === 'changing'
-          || (hasBeenValidated && revalidateOn === 'changing')
+          validateOn === "changing" ||
+          (hasBeenValidated && revalidateOn === "changing")
         ) {
           validateForm()
             .then((updatedErrors) => {
@@ -235,26 +250,26 @@ export default function useReactoForm(props) {
       if (value === null && nullValue !== undefined) value = nullValue;
 
       const inputProps = {
-        [propNames.errors || 'errors']: fieldErrors,
-        [propNames.hasBeenValidated || 'hasBeenValidated']: hasBeenValidated,
-        [propNames.isReadOnly || 'isReadOnly']: isReadOnly,
-        [propNames.name || 'name']: fieldPath,
-        [propNames.onChange || 'onChange']: onInputValueChange,
-        [propNames.onChanging || 'onChanging']: onInputValueChanging,
-        [propNames.onSubmit || 'onSubmit']: submitForm,
-        [propNames.value || 'value']: value,
+        [propNames.errors || "errors"]: fieldErrors,
+        [propNames.hasBeenValidated || "hasBeenValidated"]: hasBeenValidated,
+        [propNames.isReadOnly || "isReadOnly"]: isReadOnly,
+        [propNames.name || "name"]: fieldPath,
+        [propNames.onChange || "onChange"]: onInputValueChange,
+        [propNames.onChanging || "onChanging"]: onInputValueChanging,
+        [propNames.onSubmit || "onSubmit"]: submitForm,
+        [propNames.value || "value"]: value,
       };
 
       // If propNames key is set to `false`, omit the prop
       [
-        'errors',
-        'hasBeenValidated',
-        'isReadOnly',
-        'name',
-        'onChange',
-        'onChanging',
-        'onSubmit',
-        'value',
+        "errors",
+        "hasBeenValidated",
+        "isReadOnly",
+        "name",
+        "onChange",
+        "onChanging",
+        "onSubmit",
+        "value",
       ].forEach((key) => {
         if (propNames[key] === false) delete inputProps[key];
       });
